wchar default

+++ CommenSense/Builder/TypeBuilder.cs
@@ -21,7 +21,7 @@ partial class Builder
 		{
 			"void" => Type.Void,
 			"bool" => Type.Int1,
-			"char" => Type.Int8,
+			"char" => Type.Int16,
 			"byte" => Type.Int8,
 			"int" => Type.Int32,
 			"long" => Type.Int64,
